requirements.txt
streamlit
llama-index
cassandra-driver=
openai
boto3
python-dotenv